change type for amount (int -> float32)

+++ internal/app/dto/order.go
@@ -6,6 +6,6 @@ type Order struct {
 	Num      string    `json:"number"`
 	UserID   int       `json:"-"`
 	Status   string    `json:"status"`
-	Accrual  int       `json:"accrual"`
+	Accrual  float32   `json:"accrual"`
 	UploadAt time.Time `json:"upload_at"`
 }
